Updating license workflow to use GitHub App
For repositories with branch protection on, it is not possible to simply allow
Actions to commit changes directly to the default branch.

These changes switch to using a GitHub App for elevated permissions and open a
pull request instead.  It might be doable to allow the GitHub App to bypass
repository rules / branch protections without the need for a PR.